Проектная работа 14- в процессе работы

+++ controllers/cards.js
@@ -25,8 +25,12 @@ const createCards = (req, res) => {
 };
 
 const deleteCard = (req, res) => {
+  const owner = req.user._id;
   if (!mongoose.Types.ObjectId.isValid(req.params.cardId)) {
     return res.status(400).send({ message: 'Некорректный ID' });
+  }
+  if (owner) {
+    return res.status(400).send({ message: 'чужая карточка' });
   }
   return card.findByIdAndRemove(req.params.cardId)
     .then((newCard) => {
